Revert "Pyinstaller"
