+++ .gitignore
@@ -1,4 +1,4 @@
-*__pycache__
+/__pycache__
 /_SC4MP*
 /_Server*
 /_DMR
@@ -8,16 +8,14 @@
 /dist32
 /dist64
 /logs
-/setupbuilds
-/temp
 dbpf.py.old
 testing.py
 test_console.py
 spawn_test_servers.py
 *.bat
 *.cfg
+*/__pycache__
 *.bak
 *.log
 *.sc4
 clear_logfiles.py
-version.rc
