+++ readme.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
     <head>
-		<title>Readme &bull; SC4MP Server v0.5.0</title>
+		<title>Readme &bull; SC4MP Server v0.4.4</title>
 		<!--<link rel="stylesheet" href="resources/style.css">-->
 		<style>
 			body {
@@ -14,7 +14,7 @@
     <body>
         <center>
             <h1>SC4MP Server</h1>
-            <h4>Readme &bull; Version 0.5.0</h4>
+            <h4>Readme &bull; Version 0.4.4</h4>
             <br>
             <p>Thank you for downloading SC4MP Server! This readme will help you get started setting up your server. It's recommended that you be already familiar with the SC4MP Launcher before proeceeding.</p>
             <br>
@@ -81,7 +81,7 @@
         <h2>Updates</h2>
             <p>The latest releases of the SC4MP Server can be found on <a href="https://github.com/kegsmr/sc4mp-server/releases">GitHub</a>.</p>
         <h2>Credits</h2>
-            <p>This readme was prepared for the SC4MP Server version 0.5.0.</p>
+            <p>This readme was prepared for the SC4MP Server version 0.4.4.</p>
             <p>View the full list of collaborators on <a href="https://github.com/kegsmr/sc4mp-server/collaborators">GitHub</a>.</p>
             <p>Special thanks to <a href="http://j5.io/">Julian Cizmic</a>, the original creator of PMR, the primary inspiration for SC4MP.</p>
     </body>
